# openapi/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "maas",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "/mutex"
    }
  ],
  "paths": {
    "/lock/{name}": {
      "get": {
        "description": "Returns if a mutex with the given name is locked\n\nReturns a struct with the `name` and the `is_locked` property of a mutex.\n\n# Arguments\n\n* `name` - The name of the mutex.",
        "operationId": "is_locked",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MutexData"
                }
              }
            }
          }
        }
      },
      "put": {
        "description": "Tries to get ownership of a mutex\n\nTries to get ownership of a mutex with the specified `name`, waiting for a max amount of time specified in `timout`. This method will not wait longer than the maximum amount of time a mutex can be active. This is configured by the server owner.\n\n# Arguments\n\n* `name` - The name of the mutex.\n\n* `timeout` - The maximum amount of seconds to wait for the mutex. Default: 60. Use 0 to return instantly.\n\nThis function returns an uuid, which is proof that you hold the lock. This uuid is needed to unlock the mutex.",
        "operationId": "lock",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "timeout",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "uint64",
              "minimum": 0.0,
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LockMutexData"
                }
              }
            }
          },
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "description": "Releases ownership of a mutex\n\nReleases ownership of a mutex with the given `name`, if it is currently owned and the `uuid` matches.\n\n# Arguments\n\n* `name` - The name of the mutex.\n\n* `uuid` - The maximum amount of seconds to wait for the mutex. Default: 60. Use 0 to return instantly.\n\nreturns a struct with the `name` and the `is_locked` property of a mutex.",
        "operationId": "unlock",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "uuid",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MutexData"
                }
              }
            }
          },
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "LockMutexData": {
        "type": "object",
        "required": [
          "is_locked",
          "name",
          "uuid"
        ],
        "properties": {
          "uuid": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "is_locked": {
            "type": "boolean"
          }
        }
      },
      "CustomError": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          },
          "code": {
            "type": "integer",
            "format": "uint16",
            "minimum": 0.0
          }
        }
      },
      "MutexData": {
        "type": "object",
        "required": [
          "is_locked",
          "name"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "is_locked": {
            "type": "boolean"
          }
        }
      }
    }
  }
}